// src/types/types.ts
export type ProductType = {
  reference_id: string
  description: string
  amount: {
    value: string
  }
}
export type InfoType = {
  firstName: string
  lastName: string
  email: string
  phoneNumber: string
  address1: string
  address2: string
  city: string
  state: string
  zipCode: string
  country: string
}